Remove un-imported pyopenssl
This brings in an unpinned cryptography library, which
recently published 3.4.x, which drops python2.7 entirely,
moves the backend to rust.

To get it working, we have to upgrade to latest pip, which can
link/understand something called setuptools_rust during install.
But Pex vendors pip, so this forces us into total overhaul
territory.

I was down to pretty much our last hope, rewriting the netcode to
no longer depend on pyOpenSsl. But when I looked - turns out
that the library is listed as a dep but is unused (at least directly).

By deleting that BUILD file dependency and moving the requests
off of the [security] required set, I could work around the unpinned
cryptography and get a fullly built set of wheels.

Next step is forcing a working and full esolution.