Upgrade rust to 1.27 and force nightly cargo for ensure-installed.
Trait syntax broke backwards compat at 1.25, which we could have
allowed to stay broken (used only for development verification) but
with potential runtime misses.

This upgrades to 1.27 and enables the new flags to allow green
crate compiles. Entire rust engine is freshly compiled in the new
whls as of this change.