1.7.0+fs7
Includes pinning to single version of requests.

Very unfortunate since this will force us to upgrade pants
for any loosely-configured consumer of requests.

But the cryptography stack is basicallt broken for our version
of Python,pex and pip.

Unwinding that is a backbreaker, should be sometime in 2021
and I hate to think of it